Inziata aggiunta animazioni in ScoreState

+++ Game/OpenGLApp/OpenGLApp/MouseObject.h
@@ -20,8 +20,8 @@ public:
     }
 
     void Move(double xpos, double ypos) {
-        Position.x = xpos;
-        Position.y = ypos;
+        Position.x = (float) xpos;
+        Position.y = (float) ypos;
     }
 
     /*
